requirements.txt
python-decouple
pandas
numpy
matplotlib
GitPython
scikit-learn
requests